ai-services/job-matcher/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
scikit-learn==1.3.2
numpy>=1.25.0,<2.0.0
pandas==2.1.4
nltk==3.8.1
spacy==3.7.2
requests==2.31.0
python-dotenv==1.0.0
motor==3.3.2
pymongo==4.6.0
